% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EFL0_graphical.R
\name{EFL0_graphical}
\alias{EFL0_graphical}
\title{EFL0_graphical}
\usage{
EFL0_graphical(
  data_current,
  vars,
  rlambda = 1e-07,
  nlambda = 50,
  nlambda2 = 6,
  beta_diff_max = 1,
  adaptive = FALSE,
  adaptive_current = FALSE,
  adaptive_current_wbeta = NULL,
  adaptive_ridge = FALSE,
  L0 = FALSE,
  ncutB = 7,
  EFL0 = FALSE,
  EFL0_current = FALSE,
  EFL0_current_ridgeC = NULL,
  lambda2_seq_input = NULL,
  nfolds = 5,
  foldid_subj = NULL,
  precision_edges_true_current,
  seed = 123
)
}
\arguments{
\item{data_current}{input dataset. Must include a column \code{weight} and a
column for each node in \code{vars}.}

\item{vars}{vector of node names.}

\item{rlambda}{fraction of \code{lambda.max} to determine the smallest value
for \code{lambda}. The default is 0.0000001.}

\item{nlambda}{number of \code{lambda} values. Default is 50.}

\item{nlambda2}{number of \code{lambda2} values. Default is 6.}

\item{beta_diff_max}{optional maximum tolerated change in beta coefficients
from one time point to the next. This is used to determine the sequence of
candidate \code{lamdbda2} values for the L2 penalty. Default is 1.}

\item{adaptive}{logical indicator of whether the overall dynamic model
includes adaptive lasso weights for the L1 penalty. Default is \code{FALSE}.}

\item{adaptive_current}{logical indicator of whether to include adaptive
lasso weights for the L1 penalty in this time point's model. Default is
\code{FALSE}.}

\item{adaptive_current_wbeta}{vector of non-negative coefficient-specific
adaptive lasso weights for the L1 penalty. Default is \code{NULL}.}

\item{adaptive_ridge}{logical indicator of whether the adaptive lasso weights
should be estimated with ridge regression. Default is \code{FALSE}, which
means they will be estimated with OLS regression.}

\item{L0}{logical indicator for whether to perform L0 penalty via
hard-thresholding. Default is \code{TRUE}.}

\item{ncutB}{number of cuts to try for hard-thresholding. Default is 10.}

\item{EFL0}{logical indicator of whether the overall dynamic model
includes the two EFL0 penalties (L2 and L0). Default is \code{FALSE}.}

\item{EFL0_current}{logical indicator of whether to include the two EFL0
penalties (L2 and L0) in this time point's model. Default is \code{FALSE}.}

\item{EFL0_current_ridgeC}{vector of coefficient-specific constant C values
for the L2 penalty of EFL0. Default is\code{NULL}.}

\item{lambda2_seq_input}{vector of candidate \code{lambda2} values for the L2
penalty of EFL0. Default is \code{NULL}.}

\item{nfolds}{number of folds for cross-validation. Default is 5.}

\item{foldid_subj}{an optional vector of values between 1 and \code{nfolds}
specifying which fold each subject (or observation if there is only 1
subject) is in.}

\item{precision_edges_true_current}{optional tibble of true edges for
performance evaluation. Must include columnsm\code{row}, \code{col}, and
\code{edge}. Default is \code{NULL}.}

\item{seed}{randomization seed.}
}
\value{
A list containing the following:
\item{cvm}{cross-validation error.}
\item{coefs_adaptive_weights}{coefficients for adaptive lasso weights at the
next time point.}
\item{ridgeC_elastic_fuse}{final beta coefficients to be used in the EFL0
penalty at the next time point.}
\item{precision_edges}{a long dataset of precision matrix elements.}
\item{partial_edges}{a long dataset of partial correlation matrix elements.}
\item{edge_detection}{a matrix of tuning parameter values and performance
metrics.}
\item{output_steps}{a long dataset of beta coefficients after lasso (L1),
after soft-thresholding (L2), and after hard-thresholding (L0), when
applicable.}
}
\description{
Elastic Fuse L0 (EFL0) Network.
}
