% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EFL0_time_varying_graphical.R
\name{EFL0_time_varying_graphical}
\alias{EFL0_time_varying_graphical}
\title{EFL0_time_varying_graphical}
\usage{
EFL0_time_varying_graphical(
  data,
  vars,
  visits_of_interest,
  visits_adaptive,
  precision_edges_true = NULL,
  kernel = FALSE,
  nh = 9,
  h_power_low = -0.5,
  h_power_high = 1.5,
  adaptive = FALSE,
  adaptive_ridge = FALSE,
  L0 = FALSE,
  EFL0 = FALSE,
  beta_diff_max = 1,
  lambda2_seq_input = NULL,
  nfolds = 5,
  foldid_subj = NULL,
  seed = 123,
  ...
)
}
\arguments{
\item{data}{input dataset. Must include columns \code{id}, \code{s_il}, and
a column for each node in \code{vars}.}

\item{vars}{vector of node names.}

\item{visits_of_interest}{vector of times at which to estimate the network.}

\item{visits_adaptive}{order in which to estimate the time-specific networks.}

\item{precision_edges_true}{optional tibble of true edges for performance
evaluation. Must include columns \code{s_il}, \code{row}, \code{col},
\code{edge}. Default is \code{NULL}.}

\item{kernel}{logical indicator of whether to include observation-level
kernel weights. Default is \code{FALSE}.}

\item{nh}{number of candidate bandwidths h to try for kernel weights.
Default is 9.}

\item{h_power_low}{minimum power for constructing bandwidth sequence.
Default is -0.5.}

\item{h_power_high}{maximum power for constructing bandwidth sequence.
Default is 1.5.}

\item{adaptive}{logical indicator of whether to include adaptive lasso
weights for the L1 penalty. Default is \code{FALSE}.}

\item{adaptive_ridge}{logical indicator of whether the adaptive lasso weights
should be estimated with ridge regression. Default is \code{FALSE}, which
means they will be estimated with OLS regression.}

\item{L0}{logical indicator of whether to include an L0 penalty. Default is
\code{FALSE}.}

\item{EFL0}{logical indicator of whether to include the Elastic Fuse
penalties (L2 and L0). Default is \code{FALSE}.}

\item{beta_diff_max}{optional maximum tolerated change in beta coefficients
from one time point to the next. This is used to determine the sequence of
candidate \code{lamdbda2} values for the L2 penalty. Default is 1.}

\item{lambda2_seq_input}{user-inputted sequence of candidate \code{lambda2}
values. Default is \code{NULL}.}

\item{nfolds}{number of folds. Default is 5.}

\item{foldid_subj}{an optional vector of values between 1 and \code{nfolds}
specifying which fold each subject (or observation if there is only 1
subject) is in.}

\item{seed}{randomization seed}

\item{...}{additional arguments to be passed into \code{EFL0_graphical()}.}
}
\value{
A list containing the following:
\item{visit_weights}{a dataset of observation-level kernel weights.}
\item{precision_edges}{a long dataset of precision matrix elements.}
\item{partial_edges}{a long dataset of partial correlation matrix elements.}
\item{edge_detection}{a matrix of tuning parameter values and performance
metrics.}
\item{model_specific_cvm}{a matrix of model-specific cross-validation errors
for each node-specific model at each time point.}
\item{foldid}{the foldids that were used to perform cross-validation.}
\item{output_steps}{a long dataset of beta coefficients after lasso (L1),
after soft-thresholding (L2), and after hard-thresholding (L0), when
applicable.}
\item{h_selection}{a dataset of candidate bandwidths and resulting
cross-validation errors.}
}
\description{
Dynamic Elastic Fuse L0 (EFL0) Network.
}
