% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EFL0Network.R
\docType{package}
\name{EFL0Network}
\alias{EFL0Network}
\title{EFL0Network: A package for estimating an Elastic Fuse L0 (EFL0) dynamic
Gaussian graphical model}
\description{
The EFL0Network package provides three important functions:
EFL0_time_varying_graphical, EFL0_graphical and EFL0.
}
